+++ sdk/iothub/arm-iothub/src/api/privateLinkResourcesOperations/options.ts
@@ -4,7 +4,7 @@
 import type { OperationOptions } from "@azure-rest/core-client";
 
 /** Optional parameters. */
-export interface PrivateLinkResourcesListOptionalParams extends OperationOptions {}
+export interface PrivateLinkResourcesOperationsListOptionalParams extends OperationOptions {}
 
 /** Optional parameters. */
-export interface PrivateLinkResourcesGetOptionalParams extends OperationOptions {}
+export interface PrivateLinkResourcesOperationsGetOptionalParams extends OperationOptions {}