+++ sdk/iothub/arm-iothub/src/classic/index.ts
@@ -6,5 +6,5 @@ export type { IotHubOperations } from "./iotHub/index.js";
 export type { IotHubResourceOperations } from "./iotHubResource/index.js";
 export type { OperationsOperations } from "./operations/index.js";
 export type { PrivateEndpointConnectionsOperations } from "./privateEndpointConnections/index.js";
-export type { PrivateLinkResourcesOperations } from "./privateLinkResources/index.js";
+export type { PrivateLinkResourcesOperationsOperations } from "./privateLinkResourcesOperations/index.js";
 export type { ResourceProviderCommonOperations } from "./resourceProviderCommon/index.js";