+++ sdk/iothub/arm-iothub/src/iotHubClient.ts
@@ -13,8 +13,8 @@ import type { OperationsOperations } from "./classic/operations/index.js";
 import { _getOperationsOperations } from "./classic/operations/index.js";
 import type { PrivateEndpointConnectionsOperations } from "./classic/privateEndpointConnections/index.js";
 import { _getPrivateEndpointConnectionsOperations } from "./classic/privateEndpointConnections/index.js";
-import type { PrivateLinkResourcesOperations } from "./classic/privateLinkResources/index.js";
-import { _getPrivateLinkResourcesOperations } from "./classic/privateLinkResources/index.js";
+import type { PrivateLinkResourcesOperationsOperations } from "./classic/privateLinkResourcesOperations/index.js";
+import { _getPrivateLinkResourcesOperationsOperations } from "./classic/privateLinkResourcesOperations/index.js";
 import type { ResourceProviderCommonOperations } from "./classic/resourceProviderCommon/index.js";
 import { _getResourceProviderCommonOperations } from "./classic/resourceProviderCommon/index.js";
 import type { TokenCredential } from "@azure/core-auth";
@@ -58,7 +58,9 @@ export class IotHubClient {
     });
     this.pipeline = this._client.pipeline;
     this.resourceProviderCommon = _getResourceProviderCommonOperations(this._client);
-    this.privateLinkResources = _getPrivateLinkResourcesOperations(this._client);
+    this.privateLinkResourcesOperations = _getPrivateLinkResourcesOperationsOperations(
+      this._client,
+    );
     this.certificates = _getCertificatesOperations(this._client);
     this.iotHub = _getIotHubOperations(this._client);
     this.iotHubResource = _getIotHubResourceOperations(this._client);
@@ -68,8 +70,8 @@ export class IotHubClient {
 
   /** The operation groups for resourceProviderCommon */
   public readonly resourceProviderCommon: ResourceProviderCommonOperations;
-  /** The operation groups for privateLinkResources */
-  public readonly privateLinkResources: PrivateLinkResourcesOperations;
+  /** The operation groups for privateLinkResourcesOperations */
+  public readonly privateLinkResourcesOperations: PrivateLinkResourcesOperationsOperations;
   /** The operation groups for certificates */
   public readonly certificates: CertificatesOperations;
   /** The operation groups for iotHub */